refactor: align component structure with app architecture
