+++ app/components/ui/HIcon.vue
@@ -1,0 +1,29 @@
+<template>
+  <Icon
+    :icon="icon"
+    :height="height ?? sizeMap[size]"
+  />
+</template>
+
+<script setup lang="ts">
+import { Icon } from "@iconify/vue"
+
+export type HIconSize = "sx" | "sm" | "md" | "lg"
+
+export type HIconProps = {
+  icon?: string
+  height?: string
+  size?: HIconSize
+}
+
+const { size = "md" } = defineProps<HIconProps>()
+
+const sizeMap: Record<HIconSize, string> = {
+  sx: "12px",
+  sm: "16px",
+  md: "20px",
+  lg: "24px"
+}
+</script>
+
+<style scoped></style>
