--- layers/form/nuxt.config.ts
@@ -1,9 +1,0 @@
-export default defineNuxtConfig({
-  components: [
-    {
-      path: "~/layers/form/components/global",
-      pathPrefix: false,
-      global: true
-    }
-  ]
-})
